avm2: Convert TextFormat to AS
