+++ core/src/avm2/globals/flash/text.rs
@@ -3,4 +3,4 @@
 pub mod font;
 pub mod static_text;
 pub mod text_field;
-pub mod textformat;
+pub mod text_format;
