+++ core/src/avm2/globals/globals.as
@@ -239,6 +239,7 @@ include "flash/text/TextExtent.as"
 include "flash/text/TextField.as"
 include "flash/text/TextFieldAutoSize.as"
 include "flash/text/TextFieldType.as"
+include "flash/text/TextFormat.as"
 include "flash/text/TextFormatAlign.as"
 include "flash/text/TextInteractionMode.as"
 include "flash/text/TextLineMetrics.as"
